+++ src/xbt/cunit.c
@@ -493,7 +493,7 @@ static void apply_selection(char *selection) {
       }
     } else {
       int it;
-      for (it=0; it<xbt_dynar_length(_xbt_test_suites); it++) {
+      for (it=0; it< (int)xbt_dynar_length(_xbt_test_suites); it++) {
 	xbt_test_suite_t thissuite=xbt_dynar_get_as(_xbt_test_suites,it,xbt_test_suite_t);
 	if (!strcmp(suitename,thissuite->name)) {
 	  /* Do not disable the whole suite when we just want to disable a child */
@@ -507,7 +507,7 @@ static void apply_selection(char *selection) {
 	  } else { /* act on one child only */
 	    int it2_unit;
 	    /* search it, first (we won't reuse it for external loop which gets broken) */
-	    for (it2_unit=0; it2_unit<xbt_dynar_length(thissuite->units); it2_unit++) {
+	    for (it2_unit=0; it2_unit< (int)xbt_dynar_length(thissuite->units); it2_unit++) {
 	      xbt_test_unit_t thisunit=xbt_dynar_get_as(thissuite->units,it2_unit,xbt_test_unit_t);
 	      if (!strcmp(thisunit->name,unitname)) {
 		thisunit->enabled = enabling;
